+++ CHANGES.md
@@ -45,6 +45,8 @@
 
 ### v0.5.4
 ```
+- Add RL_STR_LOAD_UNDEFINED relation. 
+- Ensure sk_alloc_security does not generate provenance.
 - Ensure inode_setxattr does not generate provenance.
 - Ensure inode_getsecurity does not generate provenance.
 - Associate task_id with relations.